pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "music-discovery"
version = "0.1.0"
description = "Popularity-free music discovery engine using metric learning and user persona modeling"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "scikit-learn>=1.3",
    "pandas>=2.0",
    "numpy>=1.24",
    "pyarrow",
    "typer",
    "pyyaml",
    "scipy",
    "matplotlib",
    "seaborn",
    "umap-learn",
    "tqdm",
    "kaggle",
]

[project.scripts]
music-discovery = "music_discovery.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["music_discovery*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
